requirements.txt
Django>=1.11.0
djangorestframework>=3.6.3
django-wysiwyg-redactor>=0.5.0